refactor plots a bit
