+++ src/onResize/drawPlots/addViolinPlot.js
@@ -1,6 +1,6 @@
-import { layout, min, max, scale, svg, format, mean, median, quantile, deviation } from 'd3';
+import { layout, min, max, scale, svg } from 'd3';
 
-export default function addViolinPlot(chart, subgroup, violinColor = '#ccc7d6', precision = 0) {
+export default function addViolinPlot(chart, subgroup, violinColor = '#ccc7d6') {
     //Define histogram data.
     const histogram = layout
         .histogram()
@@ -95,35 +95,4 @@ export default function addViolinPlot(chart, subgroup, violinColor = '#ccc7d6', 
             stroke: violinColor,
             fill: 'none'
         });
-
-    //Annotate statistics.
-    const format0 = format(`.${precision + 0}f`);
-    const format1 = format(`.${precision + 1}f`);
-    const format2 = format(`.${precision + 2}f`);
-    subgroup.svg
-        .selectAll('g')
-        .append('title')
-        .text(
-            d =>
-                'N = ' +
-                subgroup.results.length +
-                '\nMin = ' +
-                min(subgroup.results) +
-                '\n5th % = ' +
-                format1(quantile(subgroup.results, 0.05)) +
-                '\nQ1 = ' +
-                format1(quantile(subgroup.results, 0.25)) +
-                '\nMedian = ' +
-                format1(median(subgroup.results)) +
-                '\nQ3 = ' +
-                format1(quantile(subgroup.results, 0.75)) +
-                '\n95th % = ' +
-                format1(quantile(subgroup.results, 0.95)) +
-                '\nMax = ' +
-                max(subgroup.results) +
-                '\nMean = ' +
-                format1(mean(subgroup.results)) +
-                '\nStDev = ' +
-                format2(deviation(subgroup.results))
-        );
 }
